% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_T_RPA_DETAIL.R
\name{audit_detail_fromDS1_ADJ}
\alias{audit_detail_fromDS1_ADJ}
\title{管理报表反查数据源手调凭证}
\usage{
audit_detail_fromDS1_ADJ(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FBrand = "自然堂",
  FChannel = "商超",
  FRptItemNumber = "I21"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FRptItemNumber}{报表项目代码}
}
\value{
返回值
}
\description{
管理报表反查数据源手调凭证
}
\examples{
audit_detail_fromDS1_ADJ()
}
