% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_T_FI_RPA.R
\name{audit_fi_rpa_brandChannel}
\alias{audit_fi_rpa_brandChannel}
\title{获取管报结果表的品牌渠道}
\usage{
audit_fi_rpa_brandChannel(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年}

\item{FPeriod}{月}
}
\value{
返回值
}
\description{
获取管报结果表的品牌渠道
}
\examples{
audit_fi_rpa_brandChannel()
}
