% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_T_FI_RPA.R
\name{audit_fi_rpa_rpt}
\alias{audit_fi_rpa_rpt}
\title{审计报表}
\usage{
audit_fi_rpa_rpt(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FBrand = "美素",
  FChannel = "电商",
  FOnlyError = FALSE,
  FErrorValue = 0.1
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FOnlyError}{是否只显示差异}

\item{FErrorValue}{允许的差异值，针对差异取绝对值}
}
\value{
返回值
}
\description{
审计报表
}
\examples{
audit_fi_rpa_rpt()
}
