% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_T_FI_RPA.R
\name{audit_fi_rpa_setValue}
\alias{audit_fi_rpa_setValue}
\title{审计报表进行备注查询}
\usage{
audit_fi_rpa_setValue(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FBrand = "COMO",
  FChannel = "电商",
  FRptItemNumber = "I01",
  FRptAmt = 1.08,
  FRemark = "S009方案的BW报表缺少F30客户维度，RPA重新取数"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FBrand}{品牌}

\item{FChannel}{渠道}
}
\value{
返回值
}
\description{
审计报表进行备注查询
}
\examples{
audit_fi_rpa_getValue()
}
