% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_rule_bw.R
\name{audit_rule_bw}
\alias{audit_rule_bw}
\title{反查BW规则表}
\usage{
audit_rule_bw(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 9,
  FBrand = "珀芙研",
  FChannel = "药房",
  FRptItemNumber = "I04"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FRptItemNumber}{报表项目代码}
}
\value{
返回值
}
\description{
反查BW规则表
}
\examples{
audit_rule_bw()
}
