% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_md.R
\name{md_costCenter_audit_detail}
\alias{md_costCenter_audit_detail}
\title{成本中心审计的明细项}
\usage{
md_costCenter_audit_detail(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FFinancialPartner = "周彬,李娟",
  FType = "渠道"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FFinancialPartner}{财务伙伴}

\item{FType}{渠道类型}
}
\value{
返回值
}
\description{
成本中心审计的明细项
}
\examples{
md_costCenter_audit_detail()
}
