% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_md.R
\name{md_costCenter_audit_detail_owned}
\alias{md_costCenter_audit_detail_owned}
\title{成本中心审计的明细项-独立成本中心}
\usage{
md_costCenter_audit_detail_owned(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FBrand2 = "自然堂",
  FChannel2 = "商超"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FBrand2}{品牌}

\item{FChannel2}{渠道}
}
\value{
返回值
}
\description{
成本中心审计的明细项-独立成本中心
}
\examples{
md_costCenter_audit_detail()
}
