% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_md.R
\name{md_costCenter_audit_detail_shared_list}
\alias{md_costCenter_audit_detail_shared_list}
\title{查看共享成本中心的详细信息}
\usage{
md_costCenter_audit_detail_shared_list(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FCostCenter = "20000634"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年}

\item{FPeriod}{月}

\item{FCostCenter}{成本中心}
}
\value{
返回值
}
\description{
查看共享成本中心的详细信息
}
\examples{
md_costCenter_audit_detail_shared_list()
}
