% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_md.R
\name{md_costCenter_audit_summary}
\alias{md_costCenter_audit_summary}
\title{成本中心审计}
\usage{
md_costCenter_audit_summary(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
成本中心审计
}
\examples{
md_costCenter_audit_summary()
}
