% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_md.R
\name{md_costItem_audit_summary}
\alias{md_costItem_audit_summary}
\title{检验成本要素的正确性}
\usage{
md_costItem_audit_summary(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
检验成本要素的正确性
}
\examples{
md_costItem_audit_summary()
}
