% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_channel.R
\name{mrpt2_md_costCenter_delete}
\alias{mrpt2_md_costCenter_delete}
\title{针对基础资料凭证进行重分类}
\usage{
mrpt2_md_costCenter_delete(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
针对基础资料凭证进行重分类
}
\examples{
mrpt2_md_costCenter_delete()
}
