% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_T_RPA_DETAIL.R
\name{traceBack_SAP_CostElement}
\alias{traceBack_SAP_CostElement}
\title{针对SAP数据源反向查询到相关的成本要素}
\usage{
traceBack_SAP_CostElement(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FBrand = "美素",
  FChannel = "电商",
  FRptItemNumber = "I17"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{FRptItemNumber}{报表项目代码}
}
\value{
返回值
}
\description{
针对SAP数据源反向查询到相关的成本要素
}
\examples{
traceBack_SAP_CostElement()
}
