% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_voucher.R
\name{voucher_afterReClass_costItem_summary}
\alias{voucher_afterReClass_costItem_summary}
\title{凭证重分类后按月查询凭证汇总情况}
\usage{
voucher_afterReClass_costItem_summary(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年月}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
凭证重分类后按月查询凭证汇总情况
}
\examples{
voucher_afterReClass_costItem_summary()
}
