% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_voucher.R
\name{voucher_afterReClass_vchNo_list}
\alias{voucher_afterReClass_vchNo_list}
\title{凭证号查询重分类凭证}
\usage{
voucher_afterReClass_vchNo_list(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FVchNo = "4901907219"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{FVchNo}{凭证叼}
}
\value{
返回值
}
\description{
凭证号查询重分类凭证
}
\examples{
voucher_afterReClass_vchNo_list()
}
