% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_voucher.R
\name{voucher_onlyReClass_costItem_detail_costCenterNo}
\alias{voucher_onlyReClass_costItem_detail_costCenterNo}
\title{凭证重分类后按月查询凭证展开到科目+成本中心-仅重分类凭证}
\usage{
voucher_onlyReClass_costItem_detail_costCenterNo(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FReClassifiedNumber = "D001"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年月}

\item{FPeriod}{月份}

\item{FCostItemNumber}{总账科目代码/成本要素代码}
}
\value{
返回值
}
\description{
凭证重分类后按月查询凭证展开到科目+成本中心-仅重分类凭证
}
\examples{
voucher_onlyReClass_costItem_detail_costCenterNo()
}
