% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_voucher.R
\name{voucher_onlyReClass_costItem_detail_list}
\alias{voucher_onlyReClass_costItem_detail_list}
\title{凭证重分类后按月查询凭证展开到科目+成本中心}
\usage{
voucher_onlyReClass_costItem_detail_list(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  FReClassifiedNumber = "D001",
  FCostCenterNo = "10000775"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年月}

\item{FPeriod}{月份}

\item{FCostCenterNo}{成本中心}

\item{FCostItemNumber}{总账科目代码/成本要素代码}
}
\value{
返回值
}
\description{
凭证重分类后按月查询凭证展开到科目+成本中心
}
\examples{
voucher_onlyReClass_costItem_detail_list()
}
